Added Mapping Form to UI Interactions
Added the Mapping Form to the UI interactions.  Setup the initial
DialogResult.OK for when the ok button is clicked.

+++ GBDX/VectorUpload.cs
@@ -8,6 +8,8 @@ namespace Gbdx
 {
     using System.Windows.Forms;
 
+    using Gbdx.Utilities_and_Configuration.Forms;
+
     using Ionic.Zip;
 
     public class VectorUpload : ESRI.ArcGIS.Desktop.AddIns.Button
@@ -29,35 +31,43 @@ namespace Gbdx
                     var newZip = directoryPath + "\\"
                                        + (long)(DateTime.Now - new DateTime(1970, 1, 1)).TotalMilliseconds+".zip";
 
-                    string mappingProps = "mapping.properties";
-                    const string UserContributions = "User Contributions";
-                    string itemType = "";
-                    string spatialReference = "";
+                    MappingForm mapForm = new MappingForm();
 
-                    // Write the mapping.properties file.
-                    if (!File.Exists(mappingProps))
+                    if (mapForm.ShowDialog() == DialogResult.OK)
                     {
-                        using (var sw = File.CreateText(mappingProps))
+                        string mappingProps = "mapping.properties";
+                        const string UserContributions = "User Contributions";
+                        string itemType = mapForm.ItemName;
+                        string spatialReference = "";
+
+                        // Write the mapping.properties file.
+                        if (!File.Exists(mappingProps))
                         {
-                            sw.WriteLine("vector.crs={0}", spatialReference);
-                            sw.WriteLine("vector.ingestSource={0}", UserContributions);
-                            sw.WriteLine("vector.itemType={0}", itemType);
-                            sw.WriteLine("vector.index=vector-{0}-{1}-{2}",UserContributions,itemType , DateTime.Now.ToString("yyyy-MM-dd'T'HH:mm:ss.fff'Z'"));
-                            sw.WriteLine("tagger_id=source");
-                            sw.WriteLine("id=name");
-                            sw.Flush();
-                            sw.Close();
+                            using (var sw = File.CreateText(mappingProps))
+                            {
+                                sw.WriteLine("vector.crs={0}", spatialReference);
+                                sw.WriteLine("vector.ingestSource={0}", UserContributions);
+                                sw.WriteLine("vector.itemType={0}", itemType);
+                                sw.WriteLine(
+                                    "vector.index=vector-{0}-{1}-{2}",
+                                    UserContributions,
+                                    itemType,
+                                    DateTime.Now.ToString("yyyy-MM-dd'T'HH:mm:ss.fff'Z'"));
+                                sw.WriteLine("tagger_id=source");
+                                sw.WriteLine("id=name");
+                                sw.Flush();
+                                sw.Close();
+                            }
                         }
-                    }
+                        using (var zip = new ZipFile())
+                        {
+                            zip.AddFile(openFileDialog.FileName, "");
+                            zip.AddFile(mappingProps);
+                            zip.Save(newZip);
+                        }
 
-                    using (var zip = new ZipFile())
-                    {
-                        zip.AddFile(openFileDialog.FileName,"");
-                        zip.AddFile(mappingProps);
-                        zip.Save(newZip);
+                        File.Delete(mappingProps);
                     }
-
-                    File.Delete(mappingProps);
                 }
             }
             catch (Exception error)
